+++ NEWS.md
@@ -34,5 +34,9 @@
 
 # 2018-07-30 - popkin 1.0.6.9000
 
-* internal function printLabs (used by plotPopkin) is now more flexible in where it places its labels (new args "side1" and "side2")
+* Internal function printLabs (used by plotPopkin) is now more flexible in where it places its labels (new args "side1" and "side2")
+
+# 2018-08-08 - popkin 1.0.7.9000
+
+* Added option for continuous colors, off by default.  Default is still to use only the 17 colors given directly by RColorBrewer.